% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.convolutional.R
\name{Cropping}
\alias{Cropping}
\alias{Cropping1D}
\alias{Cropping2D}
\alias{Cropping3D}
\title{Cropping layers for 1D input (e.g. temporal sequence).}
\usage{
Cropping1D(cropping = c(1, 1), input_shape = NULL)

Cropping2D(cropping = 0, data_format = NULL, input_shape = NULL)

Cropping3D(cropping = 0, data_format = NULL, input_shape = NULL)
}
\arguments{
\item{cropping}{integer or pair of integers. How many units
should be trimmed off at the beginning and end
of the cropping dimension (axis 1). If a single value
is provided, the same value will be used for both.}

\item{input_shape}{only need when first layer of a model; sets the
input shape of the data}

\item{data_format}{A string, one of channels_last (default) or
channels_first.}
}
\description{
It crops along the time dimension (axis 1).
}
\references{
Chollet, Francois. 2015. \href{https://keras.io/}{Keras: Deep Learning library for Theano and TensorFlow}.
}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
