% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callbacks.R
\name{CSVLogger}
\alias{CSVLogger}
\title{Callback that streams epoch results to a csv file.}
\usage{
CSVLogger(filename, separator = ",", append = FALSE)
}
\arguments{
\item{filename}{filename of the csv file, e.g. 'run/log.csv'.}

\item{separator}{string used to separate elements in the csv file.}

\item{append}{True: append if file exists (useful for continuing training).
False: overwrite existing file,}
}
\description{
Supports all values that can be represented as a string, including 1D
iterables such as np.ndarray.
}
\references{
Chollet, Francois. 2015. \href{https://keras.io/}{Keras: Deep Learning library for Theano and TensorFlow}.
}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
