% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applications.R
\name{decode_predictions}
\alias{decode_predictions}
\title{Decode predictions from pre-defined imagenet networks}
\usage{
decode_predictions(pred, model = c("Xception", "VGG16", "VGG19", "ResNet50",
  "InceptionV3"), top = 5)
}
\arguments{
\item{pred}{the output of predictions from the specified model}

\item{model}{the model you wish to preprocess to}

\item{top}{integer, how many top-guesses to return.}
}
\description{
These map the class integers to the actual class names
in the pre-defined models.
}
\references{
Chollet, Francois. 2015. \href{https://keras.io/}{Keras: Deep Learning library for Theano and TensorFlow}.
}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
