% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{expand_dims}
\alias{expand_dims}
\title{Expand dimensions of an array}
\usage{
expand_dims(a, axis = 0)
}
\arguments{
\item{a}{array to expand}

\item{axis}{position (amongst axes) where new axis is to be inserted.}
}
\description{
Expand the shape of an array by inserting a new axis, corresponding to
a given position in the array shape. Useful when predicting a model based
on a single input.
}
\references{
Chollet, Francois. 2015. \href{https://keras.io/}{Keras: Deep Learning library for Theano and TensorFlow}.
}
\seealso{
Other preprocessing: \code{\link{Tokenizer}},
  \code{\link{img_to_array}}, \code{\link{load_img}},
  \code{\link{one_hot}}, \code{\link{pad_sequences}},
  \code{\link{text_to_word_sequence}}
}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
