% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.pooling.R
\name{GlobalPooling}
\alias{GlobalPooling}
\alias{GlobalMaxPooling1D}
\alias{GlobalAveragePooling1D}
\alias{GlobalMaxPooling2D}
\alias{GlobalAveragePooling2D}
\title{Global pooling operations}
\usage{
GlobalMaxPooling1D(input_shape = NULL)

GlobalAveragePooling1D(input_shape = NULL)

GlobalMaxPooling2D(data_format = NULL, input_shape = NULL)

GlobalAveragePooling2D(data_format = NULL, input_shape = NULL)
}
\arguments{
\item{input_shape}{nD tensor with shape: \code{(batch_size, ..., input_dim)}.
The most common situation would be a 2D
input with shape
\code{(batch_size, input_dim)}.}

\item{data_format}{A string, one of channels_last (default)
or channels_first}
}
\description{
Global pooling operations
}
\references{
Chollet, Francois. 2015. \href{https://keras.io/}{Keras: Deep Learning library for Theano and TensorFlow}.
}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
