% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras_init.R
\name{keras_available}
\alias{keras_available}
\title{Tests if keras is available on the system.}
\usage{
keras_available(silent = FALSE)
}
\arguments{
\item{silent}{logical. Should warning message be displayed
if the result is false.}
}
\value{
Logical
}
\description{
Returns TRUE if the python \code{keras} library is installed. If the function
returns FALSE, but you believe \code{keras} is installed, then see
\link{use_python} to configure the python environment, and then
try running \link{keras_init} to establish the connection to \code{keras}.
}
\seealso{
\link{keras_init}
}
