% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{load_img}
\alias{load_img}
\title{Load image from a file as PIL object}
\usage{
load_img(path, grayscale = FALSE, target_size = NULL)
}
\arguments{
\item{path}{Path to image file}

\item{grayscale}{Boolean, whether to load the image as grayscale.}

\item{target_size}{If NULL, the default, loads the image in its native
resolution. Otherwise, set this to a vector giving
desired (img_height, img_width).}
}
\description{
Load image from a file as PIL object
}
\references{
Chollet, Francois. 2015. \href{https://keras.io/}{Keras: Deep Learning library for Theano and TensorFlow}.
}
\seealso{
Other image: \code{\link{img_to_array}}

Other preprocessing: \code{\link{Tokenizer}},
  \code{\link{expand_dims}}, \code{\link{img_to_array}},
  \code{\link{one_hot}}, \code{\link{pad_sequences}},
  \code{\link{text_to_word_sequence}}
}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
