% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.core.R
\name{Permute}
\alias{Permute}
\title{Permutes the dimensions of the input according to a given pattern.}
\usage{
Permute(dims, input_shape = NULL)
}
\arguments{
\item{dims}{vector of integers. Permutation pattern, does not include the
samples dimension. Indexing starts at 1.}

\item{input_shape}{only need when first layer of a model; sets the input shape
of the data}
}
\description{
Permutes the dimensions of the input according to a given pattern.
}
\references{
Chollet, Francois. 2015. \href{https://keras.io/}{Keras: Deep Learning library for Theano and TensorFlow}.
}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
