% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_model}
\alias{plot_model}
\title{Plot model architecture to a file}
\usage{
plot_model(model, to_file = "model.png", show_shapes = FALSE,
  show_layer_names = TRUE)
}
\arguments{
\item{model}{model object to plot}

\item{to_file}{output location of the plot)}

\item{show_shapes}{controls whether output shapes are shown in the
graph}

\item{show_layer_names}{controls whether layer names are shown in the
graph}
}
\description{
This function requires that you have installed graphviz and
pydot in Python.
}
\references{
Chollet, Francois. 2015. \href{https://keras.io/}{Keras: Deep Learning library for Theano and TensorFlow}.
}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
