% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applications.R
\name{preprocess_input}
\alias{preprocess_input}
\title{Preprocess input for pre-defined imagenet networks}
\usage{
preprocess_input(img, model = c("Xception", "VGG16", "VGG19", "ResNet50",
  "InceptionV3"))
}
\arguments{
\item{img}{the input image, as an array}

\item{model}{the model you wish to preprocess to}
}
\description{
These assume you have already converted images into a three channel,
224 by 224 matrix with \code{load_img} and \code{img_to_array}. The
processing differs based on the model so set the appropriate model
that you are using.
}
\references{
Chollet, Francois. 2015. \href{https://keras.io/}{Keras: Deep Learning library for Theano and TensorFlow}.
}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
