% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{run_examples}
\alias{run_examples}
\title{Should examples be run on this system}
\usage{
run_examples()
}
\description{
This function decided whether examples should be
run or not. Currently always answers FALSE to avoid
problems with checking the package on systems that
do not have keras installed.
}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
