% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{Sequential}
\alias{Sequential}
\title{Initialize sequential model}
\usage{
Sequential(layers = NULL)
}
\arguments{
\item{layers}{list of keras model layers}
}
\description{
Use this function to construct an empty model to which
layers will be added, or pass a list of layers directly
to the function. The first layer passed to a Sequential
model should have a defined input shape.
}
\examples{
require('kerasR')
if (run_examples()) {
X_train <- matrix(rnorm(100 * 10), nrow = 100)
Y_train <- to_categorical(matrix(sample(0:2, 100, TRUE), ncol = 1), 3)

mod <- Sequential()
mod$add(Dense(units = 50, input_shape = dim(X_train)[2]))
mod$add(Dropout(rate = 0.5))
mod$add(Activation("relu"))
mod$add(Dense(units = 3))
mod$add(ActivityRegularization(l1 = 1))
mod$add(Activation("softmax"))
keras_compile(mod,  loss = 'categorical_crossentropy', optimizer = RMSprop())

keras_fit(mod, X_train, Y_train, batch_size = 32, epochs = 5,
          verbose = 0, validation_split = 0.2)
}
}
\references{
Chollet, Francois. 2015. \href{https://keras.io/}{Keras: Deep Learning library for Theano and TensorFlow}.
}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
