% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_categorical}
\alias{to_categorical}
\title{Converts a class vector (integers) to binary class matrix.}
\usage{
to_categorical(y, num_classes = NULL)
}
\arguments{
\item{y}{class vector to be converted into a matrix (integers
from 0 to num_classes).}

\item{num_classes}{total number of classes. Set to \code{NULL} to autodetect
from the input.}
}
\description{
This function takes a vector or 1 column matrix of class labels
and converts it into a matrix with p columns, one for each category.
This is the format most commonly used in the fitting and predicting
of neural networks.
}
\references{
Chollet, Francois. 2015. \href{https://keras.io/}{Keras: Deep Learning library for Theano and TensorFlow}.
}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
