% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.convolutional.R
\name{UpSampling}
\alias{UpSampling}
\alias{UpSampling1D}
\alias{UpSampling2D}
\alias{UpSampling3D}
\title{UpSampling layers.}
\usage{
UpSampling1D(size = 2, input_shape = NULL)

UpSampling2D(size = c(2, 2), data_format = NULL, input_shape = NULL)

UpSampling3D(size = c(2, 2, 2), data_format = NULL, input_shape = NULL)
}
\arguments{
\item{size}{integer. Upsampling factor.}

\item{input_shape}{only need when first layer of a model;
sets the input shape of the data}

\item{data_format}{A string, one of channels_last
(default) or channels_first.}
}
\description{
Repeats each temporal step size a given number of times.
}
\references{
Chollet, Francois. 2015. \href{https://keras.io/}{Keras: Deep Learning library for Theano and TensorFlow}.
}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
