% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.convolutional.R
\name{ZeroPadding}
\alias{ZeroPadding}
\alias{ZeroPadding1D}
\alias{ZeroPadding2D}
\alias{ZeroPadding3D}
\title{Zero-padding layers}
\usage{
ZeroPadding1D(padding = 1, input_shape = NULL)

ZeroPadding2D(padding = 1, data_format = NULL, input_shape = NULL)

ZeroPadding3D(padding = 1, data_format = NULL, input_shape = NULL)
}
\arguments{
\item{padding}{if one integer, the same symmetric padding is applied
to width and height. If two, how many to add
for height and width.}

\item{input_shape}{only need when first layer of a model; sets the
input shape of the data}

\item{data_format}{A string, one of channels_last (default) or
channels_first.}
}
\description{
Zero-padding layers
}
\references{
Chollet, Francois. 2015. \href{https://keras.io/}{Keras: Deep Learning library for Theano and TensorFlow}.
}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
