% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.core.R
\name{Reshape}
\alias{Reshape}
\title{Reshapes an output to a certain shape.}
\usage{
Reshape(target_shape, input_shape = NULL)
}
\arguments{
\item{target_shape}{target shape. Tuple of integers, does not include
the samples dimension (batch size).}

\item{input_shape}{only need when first layer of a model; sets the input shape
of the data}
}
\description{
Reshapes an output to a certain shape.
}
\references{
Chollet, Francois. 2015. \href{https://keras.io/}{Keras: Deep Learning library for Theano and TensorFlow}.
}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
