% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{run_examples}
\alias{run_examples}
\title{Should examples be run on this system}
\usage{
run_examples()
}
\description{
This function decides whether examples should be
run or not. Answers \code{TRUE} if and only if the
package is able to find an installation of keras.
}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
