% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{text_to_word_sequence}
\alias{text_to_word_sequence}
\title{Split a sentence into a list of words.}
\usage{
text_to_word_sequence(text,
  filters = "!\\"#$\%&()*+,-./:;<=>?@[\\\\]^_`{|}~\\t\\n", lower = TRUE,
  split = " ")
}
\arguments{
\item{text}{a string}

\item{filters}{vector (or concatenation) of characters to filter out, such as punctuation.}

\item{lower}{boolean. Whether to set the text to lowercase.}

\item{split}{string. Separator for word splitting.}
}
\description{
Split a sentence into a list of words.
}
\references{
Chollet, Francois. 2015. \href{https://keras.io/}{Keras: Deep Learning library for Theano and TensorFlow}.
}
\author{
Taylor B. Arnold, \email{taylor.arnold@acm.org}
}
